F32-QC4W AVX512 GEMM microkernels using _mm256_cvtepi8_epi32
- Add 1x32 to 8x32 AVX 512 microkernels
- Load 8 bit values to 32 bit elements.
- Use set1 for params from avx params

PiperOrigin-RevId: 554552043

+++ src/microparams-init.c
@@ -2001,7 +2001,7 @@ size_t xnn_init_f32_qc4w_minmax_avx_params(
   assert(kernel_zero_point <= 15);
   params->avx.min = output_min;
   params->avx.max = output_max;
-  params->avx.kernel_zero_point = (int32_t) kernel_zero_point;
+  params->avx.kernel_zero_point = kernel_zero_point;
   params->avx.mask = 0xF;
   return sizeof(params->avx);
 }
